Main.MoMForward.find_grids_with_object — Method

find_grids_with_object(config, scatter; bacground_permittivity = 1.0)

We find the position in the matrix and the of coordinates the scatter grid, whose complex permittivity is not equal to that of the background.

Arguments

config::ConstantParameter: The structure of the configuration.

scatter::Array: The permittivity of the object and background, which is a 2D array of booleans.

Keyword Arguments

bacground_permittivity::Float64: The permittivity of the background, which is 1.0 by default.

Returns

scatter_positions_in_matrix::Array: The position in the matrix of the scatter grid.

scatter_centroids::Array: The coordinates of the scatter grid

Main.MoMForward.ConstantParameter — Type

ConstantParameter(;
freq,
doi_size_x,
doi_size_y,
grid_number_x,
grid_number_y,
R_tx,
R_rx,
tx_num,
rxs_num
)

Construct a structure based on the frequency of the input, including the following parameters:

ϵ₀ is the permittivity of free space.

μ₀ is the permeability of free space.

c is the speed of light in free space.

λ is the wavelength.

f is the frequency.

ω is the angular frequency.

k₀ is the wave number in the vacuum.

η is the intrinsic impedance of free space.

constant1 is a constant factor, which is μ₀ / (4 * pi).

constant2 is a constant factor, which is -1im / (4 * pi * ω * ϵ₀).

doi_size_x is the size of the DOI in the x direction.

doi_size_y is the size of the DOI in the y direction.

grid_number_x is the number of grids in the x direction.

grid_number_y is the number of grids in the y direction.

grid_num is the total number of grids, which is grid_number_x * grid_number_y.

grid_radius is the radius of a circle with the same area as the grid.

grid_xs is the x coordinates of the grids.

grid_ys is the y coordinates of the grids.

txs_pos is the position of the transmitters, whose size is [2, tx_num].

rxs_pos is the position of the receivers, whose size is [2, rxs_num].

tx_num is the number of transmitters.

rxs_num is the number of receivers.

Keyword Arguments

freq::Float64: The frequency of the input.

doi_size_x::Float64: The size of the DOI in the x direction.

doi_size_y::Float64: The size of the DOI in the y direction

grid_number_x::Int64: The number of grids in the x direction.

grid_number_y::Int64: The number of grids in the y direction.

R_tx: The radius of the transmitter circle.

R_rx: The radius of the receiver circle.

tx_num: The number of transmitters.

rxs_num: The number of receivers.