Main.PreProcessing.getLinearEquation — Method

getLinearEquation(p1x, p1y, p2x, p2y)

Returns the coefficients of a straight line.

Arguments:

p1x: The x coordinate of the first point on the line.

p1y: The y coordinate of the first point on the line.

p2x: The x coordinate of the second point on the line.

p2y: The y coordinate of the second point on the line.

Returns:

The coefficients of the straight line, $[a, b, c]$, where $ax + by + c = 0$.

Main.PreProcessing.get_csiabs2_and_scatter — Method

get_csi_and_label(material_names, material_labels, rectangle_pos, folder,parameters::ConstantParameter,selected_freq_index; tx_num=4, rx_num=40,sample_num=10,window_size=100)

Read the label from the xlsx file and get the CSI data. The size of CSI data is (rxnum, txnum, freqnum, samplenum).

Main.PreProcessing.get_dis_between_a_point_and_a_line — Method

get_dis_between_a_point_and_a_line(point, line_point1, line_point2)

Returns the distance between a point and a straight line. With a line is $ax + by + c = 0$, the distance is $|ax + by + c| \sqrt{a^2 + b^2}$.

Arguments:

point: The coordinates of the point, $[x, y]$.

line_point1: The coordinates of the first point on the line, $[x, y]$.

line_point2: The coordinates of the second point on the line, $[x, y]$.

line_point1 and line_point2 is used to calculate straight-line equations, getLinearEquation(p1x, p1y, p2x, p2y).

Returns:

The distance between the point and the line.